Initial Commit
Thesis project, with different back-ups and datasets used, datasets used in experiments are in the data folder, while others which could help are currently in datasets. In the code folder there are the files of the experiments, while in notes there are some annotations I have done to follow the process while writing the memory.